Fix - I'm nele: Won't tag someone who's already in vc

+++ src/functions/im-nele/main.js
@@ -6,13 +6,16 @@ const { nele, royalUserIds } = Config;
 export const init = async (client) => {
     const { voiceId, textId } = nele;
     const textChannel = client.channels.cache.get(textId);
+    const channel = client.channels.cache.get(voiceId);
     client.on('voiceStateUpdate', async (prev, curr) => {
         if (curr.channelId !== voiceId) return;
         if (prev.channelId === curr.channelId) return;
         const index = royalUserIds.indexOf(curr.id);
         if (index === -1) return;
-        const target = royalUserIds[1 - index];
-        await wipeChannel(textChannel);
-        textChannel.send(`<@${target}> I'm nele`);
+        const userId = royalUserIds[1 - index];
+        if (!channel?.members.get(userId)) {
+            await wipeChannel(textChannel);
+            textChannel.send(`<@${userId}> I'm nele`);
+        }
     });
 };
